+++ bootldr/dynlink.c
@@ -25,6 +25,10 @@ static hashtab_t loaded_libs;
 // 	return (size_t)h;
 // }
 
+#define ALLOC_KPAGES(ptr, size) uefi_call_wrapper(BS->AllocatePages, 4, AllocateAnyPages, EfiLoaderData, size, ptr)
+static inline size_t div_up(size_t a, size_t b) { return (a+b-1)/b; }
+#define TO_PAGES(ptr) div_up(ptr, 4096)
+
 #define foreach_needed(d, code) for (Elf64_Dyn *c = d; c->d_tag != DT_NULL; c++) { if (c->d_tag == DT_NEEDED) { code }}
 
 #define USE_ALLOCA
@@ -42,6 +46,8 @@ static void parse_exports(hashtab_t *const ctx, libid_t *lib){
 				.addr = (uint64_t)CONV_PTR(lib->image + lib->dynsym[i].st_value)
 			};
 			ht_insert(ctx, &lib->dynstr[lib->dynsym[i].st_name], &sym);
+		} else if (ELF64_ST_BIND(lib->dynsym[i].st_info) == STB_WEAK && !strcmpa(&lib->dynstr[lib->dynsym[i].st_name], "_DTV_")){
+			*(uint64_t*)(lib->image + lib->dynsym[i].st_value) = (uint64_t)CONV_PTR(lib->image + 0x1000);
 		}
 	}
 }
@@ -78,9 +84,16 @@ static libid_weakptr get_file(hashtab_t *const ctx, char* path){
 	//Print(L"NL rela: %p\n", nl.rela);
 	//Print(L"NL relacnt: %p\n", nl.relacnt);
 
+	nl.tls_area = NULL;
+	nl.tls_size = 0;
+
 	Elf64_Shdr* percpu = get_section_by_name(nl.image, ".percpu");
 	if (percpu){
-		*(uint64_t*)(nl.image + 0x0ff8) = (uint64_t)CONV_PTR(nl.image + percpu->sh_addr);
+		nl.tls_size = percpu->sh_size;
+		if (EFI_ERROR(ALLOC_KPAGES(&nl.tls_area, nl.tls_size))) goto fail;
+
+		CopyMem(nl.tls_area, nl.image + percpu->sh_addr, nl.tls_size);
+		*(uint64_t*)(nl.image + 0x0ff8) = (uint64_t)CONV_PTR(nl.tls_area);
 	}
 
 	foreach_needed(nl.dynamic, 
@@ -136,15 +149,11 @@ static void resolve(hashtab_t *const ctx, libid_t *lib){
 				//Print(L"GLOB_DAT\n");
 				char *symname = &lib->dynstr[lib->dynsym[ELF64_R_SYM(lib->rela[i].r_info)].st_name];
 
-				if (strcmpa(symname,"_DTV_")) {
-					symbol_t *target = ht_read(ctx, symname);
-					if (!target && ctx != &global_exports) target = ht_read(&global_exports, symname);
-					if (!target) Fatality(LDR_ELF, 24);
-					*(uint64_t*)got_entry = target->addr + lib->rela[i].r_addend;
-				} else {
-					Print(L"new _DTV_:%p\n", CONV_PTR(lib->image + 0x1000));
-					*(uint64_t*)got_entry = (uint64_t)CONV_PTR(lib->image + 0x1000);
-				}
+				symbol_t *target = ht_read(ctx, symname);
+				if (!target && ctx != &global_exports) target = ht_read(&global_exports, symname);
+				if (!target) Fatality(LDR_ELF, 24);
+				*(uint64_t*)got_entry = target->addr + lib->rela[i].r_addend;
+
 				break;
 			case R_X86_64_NONE:
 				//Print(L"NONE\n");
@@ -162,7 +171,7 @@ static void resolve(hashtab_t *const ctx, libid_t *lib){
 	)
 }
 
-libid_weakptr dl_process(char *path, const bool global_f){
+libid_weakptr dl_process(char *path, const BOOLEAN global_f){
 #ifndef USE_ALLOCA
 	hashtab_t local;
 	hashtab_t *const ctx = global_f ? &global_exports : &local;
@@ -185,10 +194,6 @@ libid_weakptr dl_process(char *path, const bool global_f){
 	return module;
 }
 
-#define ALLOC_KPAGES(ptr, size) uefi_call_wrapper(BS->AllocatePages, 4, AllocateAnyPages, EfiLoaderData, size, ptr)
-static inline size_t div_up(size_t a, size_t b) { return (a+b-1)/b; }
-#define TO_PAGES(ptr) div_up(ptr, 4096)
-
 void dl_fini(loaded_file **files, size_t *files_cnt){
 	ht_destroy(&global_exports);
 
@@ -198,10 +203,13 @@ void dl_fini(loaded_file **files, size_t *files_cnt){
 			ht_entry_t* ent = ht_index(&loaded_libs, i);
 			if (ent->key){
 				libid_t *lib = (libid_t*)((uint8_t*)ent + sizeof(ht_entry_t));
-				(*files)[*files_cnt].name = lib->soname;
-				(*files)[*files_cnt].data = CONV_PTR(lib->image);
-				(*files)[*files_cnt].size = lib->imagesize;
-				(*files_cnt)++;
+				loaded_file *file = *files + (*files_cnt)++;
+
+				file->name = lib->soname;
+				file->data = CONV_PTR(lib->image);
+				file->size = lib->imagesize;
+				file->tls_area = lib->tls_area ? CONV_PTR(lib->tls_area) : NULL;
+				file->tls_size = lib->tls_size;
 			}
 		}
 	} else {