+++ kernel/include/boot.h
@@ -1,12 +1,15 @@
 #ifndef __BOOTINFO_H__
 #define __BOOTINFO_H__
 
+#include "physmem.h"
 #include "rtsvcs.h"
 
 typedef struct __packed{
 	char* name;
 	void* data;
 	size_t size;
+	void* tls_area;
+	size_t tls_size;
 } loaded_file;
 
 typedef struct __packed{