+++ kernel/include/lib/types.h
@@ -1,30 +1,36 @@
 #ifndef __TYPES_H__
 #define __TYPES_H__
 
-typedef __INT8_TYPE__ int8_t;
-typedef __UINT8_TYPE__ uint8_t;
+#include <stdint.h>
+#include <stddef.h>
 
-typedef __INT16_TYPE__ int16_t;
-typedef __UINT16_TYPE__ uint16_t;
+// typedef __INT8_TYPE__ int8_t;
+// typedef __UINT8_TYPE__ uint8_t;
 
-typedef __INT32_TYPE__ int32_t;
-typedef __UINT32_TYPE__ uint32_t;
+// typedef __INT16_TYPE__ int16_t;
+// typedef __UINT16_TYPE__ uint16_t;
 
-typedef __INT64_TYPE__ int64_t;
-typedef __UINT64_TYPE__ uint64_t;
+// typedef __INT32_TYPE__ int32_t;
+// typedef __UINT32_TYPE__ uint32_t;
 
-typedef __SIZE_TYPE__ size_t;
-typedef __PTRDIFF_TYPE__ ptrdiff_t;
+// typedef __INT64_TYPE__ int64_t;
+// typedef __UINT64_TYPE__ uint64_t;
 
-typedef _Bool bool;
+// typedef __SIZE_TYPE__ size_t;
+// typedef __PTRDIFF_TYPE__ ptrdiff_t;
 
-#define NULL ((void*)0)
+// typedef __UINTPTR_TYPE__ uintptr_t;
+// typedef __INTPTR_TYPE__ intptr_t;
+
+// typedef _Bool bool;
+
+//#define NULL ((void*)0)
 #define true 1
 #define false 0
 
 #define asm __asm__
 
-#define __export __attribute__((used,visibility("default")))
+#define __export __attribute__((visibility("default")))
 #define __always_inline __attribute__((always_inline))
 #define __noinline __attribute__((noinline))
 #define __packed __attribute__((__packed__))